Log Node has been renamed from Queue to TaskQueue
This is to prevent confusion/conflicts with other queues (e.g. the ones from workflow). Note that this change will be effective from Mendix `9.0.4` onwards, so the log node will still be called "Queue" in Mendix `9.0.3`. This has been discussed with product management and is deemed OK, because not many `9.0.3` customers will be using it anyway.